Refactor website modules
