+++ sister_website/__init__.py
@@ -1,0 +1,1 @@
+# Package for SISTER website
